Refactoring of the whole crate
